fix: divided by 0 inside ex03, change type of toInt in int not float,

+++ ex02/inc/Fixed.hpp
@@ -18,33 +18,33 @@
 class	Fixed
 {
 	public:
-		Fixed ( void );
-		Fixed ( const Fixed& obj );
-		Fixed ( const int nb );
-		Fixed ( const float nb );
-		~Fixed ( void );
+		Fixed	( void );
+		Fixed	( const Fixed& obj );
+		Fixed	( const int nb );
+		Fixed	( const float nb );
+		~Fixed	( void );
 
-		Fixed& operator=( const Fixed &other );
-		bool operator>( const Fixed &other ) const;
-		bool operator<( const Fixed &other ) const;
-		bool operator>=( const Fixed &other ) const;
-		bool operator<=( const Fixed &other ) const;
-		bool operator!=( const Fixed &other ) const;
-		bool operator==( const Fixed &other ) const;
+		Fixed&	operator=( const Fixed &other );
+		bool	operator>( const Fixed &other ) const;
+		bool	operator<( const Fixed &other ) const;
+		bool	operator>=( const Fixed &other ) const;
+		bool	operator<=( const Fixed &other ) const;
+		bool	operator!=( const Fixed &other ) const;
+		bool	operator==( const Fixed &other ) const;
 
 		Fixed&	operator+( const Fixed &other );
 		Fixed&	operator-( const Fixed &other );
-		Fixed operator*(Fixed const &rhs) const; 
+		Fixed	operator*(Fixed const &rhs) const; 
 		Fixed&	operator/( const Fixed &other );
 		Fixed&	operator++();
 		Fixed	operator++( int );
 		Fixed&	operator--();
 		Fixed	operator--( int );
  
-		int getRawBits( void ) const;
-		void setRawBits( int const raw );
-		float toFloat( void ) const;
-		float toInt( void ) const;
+		int		getRawBits( void ) const;
+		void	setRawBits( int const raw );
+		float	toFloat( void ) const;
+		int		toInt( void ) const;
 
 		static Fixed& min(Fixed &other, Fixed &other2);
 		static const Fixed& min(const Fixed &other, const Fixed &other2);
